refactor: move auto-fill dialog from generic header to QT page
PredictionStatusHeader no longer owns the auto-fill dialog — it now
calls action.onClick directly for all onClick actions. The QT-specific
confirmation dialog and bulkAutoFillFromPredictions logic move to
qualified-teams-client-page.tsx where they belong.

This also fixes the awards page "Finalizar" button which was incorrectly
opening the auto-fill dialog instead of scrolling to the next award.
