+++ app/components/prediction-status-header/prediction-status-header.tsx
@@ -1,19 +1,11 @@
 'use client';
 
-import { useState, useTransition } from 'react';
 import {
   Box,
   Button,
   Card,
   Chip,
-  CircularProgress,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
   Divider,
-  Snackbar,
   Typography,
 } from '@mui/material';
 import { alpha, type Theme } from '@mui/material/styles';
@@ -26,18 +18,11 @@ import RocketLaunchIcon from '@mui/icons-material/RocketLaunch';
 import FlagIcon from '@mui/icons-material/Flag';
 import EmojiEventsIcon from '@mui/icons-material/EmojiEvents';
 import Link from 'next/link';
-import { useTranslations, useLocale } from 'next-intl';
-import { bulkAutoFillFromPredictions } from '../../actions/qualification-actions';
-import { toLocale } from '../../utils/locale-utils';
-import type { QualifiedTeamPrediction } from '../../db/tables-definition';
 import { BoostCountBadge } from '../boost-badge';
 import type { StatusHeaderVariant, StatusHeaderTone, HeaderAction } from './types';
 
 interface PredictionStatusHeaderProps {
   variant: StatusHeaderVariant;
-  onAutoFillSuccess?: (predictions: QualifiedTeamPrediction[]) => void;
-  isLocked?: boolean;
-  tournamentId?: string;
 }
 
 function getBgColor(theme: Theme, tone: StatusHeaderTone): string {
@@ -92,78 +77,22 @@ function LeadIcon({ icon, color }: { icon: StatusHeaderVariant['leadIcon']; colo
   }
 }
 
-function AutoFillDialog({
-  open, isCalculating, onClose, onConfirm,
-}: { open: boolean; isCalculating: boolean; onClose: () => void; onConfirm: () => void }) {
-  const t = useTranslations('qualified-teams.nudge');
-  return (
-    <Dialog open={open} onClose={isCalculating ? undefined : onClose} disableEscapeKeyDown={isCalculating}>
-      <DialogTitle>{t('autoFillDialog.title')}</DialogTitle>
-      <DialogContent>
-        <DialogContentText>{t('autoFillDialog.body')}</DialogContentText>
-        <DialogContentText sx={{ mt: 1 }}>{t('autoFillDialog.note')}</DialogContentText>
-      </DialogContent>
-      <DialogActions>
-        <Button onClick={onClose} disabled={isCalculating}>{t('autoFillDialog.cancel')}</Button>
-        <Button
-          onClick={onConfirm}
-          variant="contained"
-          color="primary"
-          disabled={isCalculating}
-          autoFocus
-          startIcon={isCalculating ? <CircularProgress size={16} color="inherit" /> : undefined}
-        >
-          {isCalculating ? t('autoFillDialog.calculating') : t('autoFillDialog.confirm')}
-        </Button>
-      </DialogActions>
-    </Dialog>
-  );
-}
-
-export function PredictionStatusHeader({
-  variant,
-  onAutoFillSuccess,
-  isLocked,
-  tournamentId,
-}: PredictionStatusHeaderProps) {
-  const locale = toLocale(useLocale());
-  const tNudge = useTranslations('qualified-teams.nudge');
-  const [dialogOpen, setDialogOpen] = useState(false);
-  const [isCalculating, setIsCalculating] = useState(false);
-  const [errorOpen, setErrorOpen] = useState(false);
-  const [, startTransition] = useTransition();
-
-  const handleAutoFillConfirm = () => {
-    if (!tournamentId) return;
-    setIsCalculating(true);
-    startTransition(async () => {
-      const result = await bulkAutoFillFromPredictions(tournamentId, locale);
-      setIsCalculating(false);
-      setDialogOpen(false);
-      if (result.success && result.predictions) {
-        onAutoFillSuccess?.(result.predictions);
-      } else {
-        setErrorOpen(true);
-      }
-    });
-  };
-
+export function PredictionStatusHeader({ variant }: PredictionStatusHeaderProps) {
   const accentColor = variant.statusColor ?? getToneColor(variant.tone);
   const borderColor = getBorderColor(variant.tone);
   const isExpanded = !!variant.message;
-  const hasOnClickAction = !!(variant.action && 'onClick' in variant.action);
 
   function renderAction(action: HeaderAction, buttonVariant: 'text' | 'contained' | 'outlined') {
     const sx = { flexShrink: 0, whiteSpace: 'nowrap', ...(buttonVariant === 'text' ? { fontSize: 12, py: 0.25, minWidth: 0 } : {}) };
     if ('href' in action) {
       return (
-        <Button size="small" variant={buttonVariant} color="primary" component={Link} href={action.href} sx={sx} disabled={isLocked}>
+        <Button size="small" variant={buttonVariant} color="primary" component={Link} href={action.href} sx={sx}>
           {action.label}
         </Button>
       );
     }
     return (
-      <Button size="small" variant={buttonVariant} color="primary" onClick={() => setDialogOpen(true)} sx={sx} disabled={isLocked}>
+      <Button size="small" variant={buttonVariant} color="primary" onClick={action.onClick} sx={sx}>
         {action.label}
       </Button>
     );
@@ -172,8 +101,7 @@ export function PredictionStatusHeader({
   const bothActions = variant.action && variant.secondaryAction;
 
   return (
-    <>
-      <Card
+    <Card
         sx={{
           bgcolor: (theme) => getBgColor(theme, variant.tone),
           border: 1,
@@ -273,22 +201,5 @@ export function PredictionStatusHeader({
           </>
         )}
       </Card>
-
-      {hasOnClickAction && (
-        <AutoFillDialog
-          open={dialogOpen}
-          isCalculating={isCalculating}
-          onClose={() => setDialogOpen(false)}
-          onConfirm={handleAutoFillConfirm}
-        />
-      )}
-
-      <Snackbar
-        open={errorOpen}
-        autoHideDuration={4000}
-        onClose={() => setErrorOpen(false)}
-        message={tNudge('autoFillError')}
-      />
-    </>
   );
 }
